🎨 moving around some functions for clarity
