+++ router/utils.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"bufio"
+)
+
+func readLn(r *bufio.Reader) (string, error) {
+	var (
+		isPrefix = true
+		err      error
+		line, ln []byte
+	)
+	for isPrefix && err == nil {
+		line, isPrefix, err = r.ReadLine()
+		ln = append(ln, line...)
+	}
+	if err != nil {
+		return "", err
+	}
+	return string(ln), nil
+}
